BusyTimer: break up the smart home code
