+++ applications/services/busy_timer/busy_timer_smart_home.c
@@ -1,0 +1,125 @@
+#include "busy_timer_i.h"
+
+#define TIMER_SMART_HOME_PROFILE_ID (BusyTimerProfileIdBusy)
+
+static void
+    busy_timer_smart_home_handle_timer_event(BusyTimer* instance, const BusyTimerEvent* event) {
+    const BusyTimerEventType event_type = event->type;
+
+    MatterSwitchState switch_state = MatterSwitchStateMax;
+
+    if(event_type == BusyTimerEventTypeStateChanged) {
+        const BusyTimerState timer_state = event->state_changed.state;
+        switch_state = (timer_state == BusyTimerStateWork) ? MatterSwitchStateOn :
+                                                             MatterSwitchStateOff;
+    } else if(event_type == BusyTimerEventTypePaused) {
+        const bool is_paused = event->paused.is_paused;
+        switch_state = is_paused ? MatterSwitchStateOff : MatterSwitchStateOn;
+
+    } else if(event_type == BusyTimerEventTypeIntervalEnded) {
+        switch_state = MatterSwitchStateOff;
+    }
+
+    if((switch_state != MatterSwitchStateMax) && (switch_state != instance->matter_switch_state)) {
+        instance->matter_switch_state = switch_state;
+        matter_set_switch_state(instance->matter, switch_state);
+    }
+}
+
+static void busy_timer_smart_home_timer_pubsub_callback(const void* message, void* context) {
+    furi_assert(message);
+    furi_assert(context);
+
+    BusyTimer* instance = context;
+    const BusyTimerEvent* event = message;
+
+    if(instance->app_config.is_smart_home_enabled) {
+        busy_timer_smart_home_handle_timer_event(instance, event);
+    }
+}
+
+static void busy_timer_matter_switch_state_callback(const void* item, void* context) {
+    furi_assert(item);
+    furi_assert(context);
+
+    BusyTimer* instance = context;
+
+    const MatterSwitchState switch_state = *(MatterSwitchState*)item;
+    furi_assert(switch_state < MatterSwitchStateMax);
+
+    busy_timer_handle_matter(instance, switch_state);
+}
+
+static void busy_timer_smart_home_start_app(BusyTimer* instance) {
+    if(instance->state == BusyTimerStateIdle) {
+        busy_timer_apply_profile_settings(instance, TIMER_SMART_HOME_PROFILE_ID);
+    }
+
+    BusyTimerSettings* settings = &instance->settings[TIMER_SMART_HOME_PROFILE_ID];
+    BusyAppConfig* app_config = &settings->profile.app_config;
+
+    app_config->is_smart_home_enabled = true;
+
+    busy_timer_start_app(app_config);
+    busy_timer_start_internal(instance);
+}
+
+void busy_timer_smart_home_init(BusyTimer* instance) {
+    instance->matter = furi_record_open(RECORD_MATTER);
+
+    furi_pubsub_subscribe(
+        instance->event_pubsub, busy_timer_smart_home_timer_pubsub_callback, instance);
+
+    furi_state_subscribe(
+        matter_get_switch_state(instance->matter),
+        busy_timer_matter_switch_state_callback,
+        instance);
+}
+
+void busy_timer_smart_home_handle_switch_state(BusyTimer* instance, MatterSwitchState switch_state) {
+    if(instance->matter_switch_state == switch_state) {
+        return;
+    }
+
+    if(switch_state == MatterSwitchStateOn) {
+        if(instance->state == BusyTimerStateIdle) {
+            busy_timer_smart_home_start_app(instance);
+
+        } else if(instance->state == BusyTimerStateWork) {
+            if(!busy_timer_is_running(instance)) {
+                if(instance->time_elapsed_s == 0) {
+                    busy_timer_smart_home_start_app(instance);
+                } else {
+                    busy_timer_toggle_internal(instance);
+                }
+            }
+
+        } else if(instance->state == BusyTimerStateRest) {
+            if(!busy_timer_is_running(instance)) {
+                if(instance->time_elapsed_s == 0) {
+                    busy_timer_smart_home_start_app(instance);
+
+                    switch_state = MatterSwitchStateOff;
+                    matter_set_switch_state(instance->matter, switch_state);
+
+                } else {
+                    busy_timer_skip_internal(instance);
+                }
+
+            } else {
+                busy_timer_skip_internal(instance);
+            }
+        }
+
+    } else if(switch_state == MatterSwitchStateOff) {
+        if(instance->state != BusyTimerStateIdle) {
+            busy_timer_stop_internal(instance);
+            busy_timer_exit_app();
+        }
+
+    } else {
+        return;
+    }
+
+    instance->matter_switch_state = switch_state;
+}
